Move jsonschema code from Python dir to raml pacakge.
